refactor: move the discovery view's payload builders into modules
The option-list builders and the node-type lookup were methods on the
discovery views, which left them unreachable from anywhere else. Both are
about the payload rather than about serving it, so they move next to the
rules they apply: the option lists into `discovery/options.py` and the
by-name lookup into `discovery/node_types.py`.
